# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: The Waffle Hub BCM - KOT & Billing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Orders ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, price, active, createdAt]
      properties:
        id:          { type: integer }
        categoryId:  { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        name:        { type: string }
        description: { type: ["string", "null"] }
        price:       { type: number }
        active:      { type: boolean }
        createdAt:   { type: string }
    ProductInput:
      type: object
      required: [name, price]
      properties:
        categoryId:  { type: ["integer", "null"] }
        name:        { type: string, minLength: 1 }
        description: { type: string }
        price:       { type: number, minimum: 0 }
        active:      { type: boolean }
    ProductUpdate:
      type: object
      properties:
        categoryId:  { type: ["integer", "null"] }
        name:        { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        price:       { type: number, minimum: 0 }
        active:      { type: boolean }

    # ── Order ─────────────────────────────────────────────────────────────────
...